Rename utils.ts to index.ts
